+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 
+## [1.1.3] - Unreleased
+- Add WidgetExt::with_type().
+- Add MenuExt::menu and set_menu().
+
 ## [1.1.2] - 2021-07-19
 - Fix app::event_clicks() and add overload for event_clicks_num() for number of clicks.
 - Add widget_resize() for simple resizing without class-specific resize functionality.